apn-app/requirements.txt
asgiref==3.2.3
astroid==2.4.1
certifi==2019.11.28
chardet==3.0.4
dict.sorted==1.0.0
Django==3.0.4
django-cleanup==4.0.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-multiselectfield==0.1.12
djangorestframework==3.11.0
drf-nested-relations==0.0.2
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.5.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
python-dotenv==0.15.0
pytils==0.3
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
urllib3==1.25.8
wrapt==1.12.1
drf-yasg==1.20.0